+++ CHANGELOG.md
@@ -51,3 +51,4 @@
 [@dsalahutdinov]: https://github.com/dsalahutdinov
 [@Envek]: https://github.com/Envek
 [@DmitryTsepelev]: https://github.com/DmitryTsepelev
+[@shivanshgaur]: https://github.com/shivanshgaur